Core Infrastructure/bookshelf/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-logging==3.6.0
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0